DefaultObject splitted
