+++ interface/src/js/defaultObjects/boss.js
@@ -1,0 +1,6 @@
+export function boss() {
+    return {
+        Name: "", QuickMove: "", ChargeMove: "",
+        Tier: "4", quickMovePool: "", chargeMovePool: "",
+    }
+}
