+++ interface/src/js/defaultObjects/getUserPok.js
@@ -1,0 +1,7 @@
+export function getUserPok() {
+    return {
+        Name: "", QuickMove: "", ChargeMove: "", ChargeMove2: "",
+        Lvl: "40", Atk: "15", Def: "15", Sta: "15",
+        IsShadow: "false", quickMovePool: "", chargeMovePool: "",
+    }
+}
