+++ interface/src/js/defaultObjects/pokemon.js
@@ -1,0 +1,15 @@
+export function pokemon() {
+    return {
+        name: "", Lvl: "", Atk: "", Def: "", Sta: "", Shields: "0",
+        AtkStage: "0", DefStage: "0", InitialHP: "0", InitialEnergy: "0", IsGreedy: "true", IsShadow: "false",
+        QuickMove: "", ChargeMove1: "", ChargeMove2: "",
+        quickMovePool: "", chargeMovePool: "",
+        effAtk: "", effDef: "", effSta: "",
+        maximizer: {
+            stat: "Overall",
+            level: "40",
+            action: "Default",
+        },
+        HP: undefined, Energy: undefined, showMenu: false,
+    }
+}
