+++ interface/src/js/defaultObjects/pveCustomParty.js
@@ -1,0 +1,3 @@
+export function pveCustomParty() {
+    return { title: "", party: [] }
+}
