+++ interface/src/js/defaultObjects/pveUserSettings.js
@@ -1,0 +1,5 @@
+import { pveCustomParty } from "./pveCustomParty";
+
+export function pveUserSettings() {
+    return { FindInCollection: true, SortByDamage: "true", UserPlayers: [[pveCustomParty(), pveCustomParty(), pveCustomParty()],], }
+}
