+++ interface/src/js/defaultObjects/pveattacker.js
@@ -1,0 +1,7 @@
+export function pveattacker() {
+    return {
+        Name: "", QuickMove: "", ChargeMove: "",
+        Lvl: "40", Atk: "15", Def: "15", Sta: "15",
+        IsShadow: "false", quickMovePool: "", chargeMovePool: "",
+    }
+}
