+++ interface/src/js/defaultObjects/pveobj.js
@@ -1,0 +1,6 @@
+export function pveobj() {
+    return {
+        FriendshipStage: "0", Weather: "0", DodgeStrategy: "0",
+        PartySize: "18", PlayersNumber: "3", IsAggresive: "true", SupportSlotEnabled: "false",
+    }
+}
