+++ interface/src/js/encoders/encodePveBoss.js
@@ -1,0 +1,4 @@
+export function encodePveBoss(data) {
+    let res = [data.Name, data.QuickMove, data.ChargeMove, data.Tier,]
+    return encodeURIComponent(res.join("_"));
+}
